little changes

+++ Server/ReasnAPI/ReasnAPI/Services/ParameterService.cs
@@ -6,15 +6,6 @@ namespace ReasnAPI.Services {
     public class ParameterService (ReasnContext context){
         private readonly ReasnContext _context = context;
 
-        /* TODO: Create following functions for this class
-         * create
-         * update
-         * delete
-         * get by ID
-         * get list by filter
-         * get all
-         */
-
         public ParameterDto CreateParameter(ParameterDto parameterDto)
         {
             var parameter = new Parameter
@@ -40,23 +31,20 @@ namespace ReasnAPI.Services {
 
             parameter.Key = parameterDto.Key;
             parameter.Value = parameterDto.Value;
-
             _context.Parameters.Update(parameter);
             _context.SaveChanges();
-
             return parameterDto;
         }
 
         public void DeleteParameter(int parameterId)
         {
             var parameter = _context.Parameters.FirstOrDefault(r => r.Id == parameterId);
-           
-      
-
+            if (parameter == null)
+            {
+                return;
+            }
             _context.Parameters.Remove(parameter);
             _context.SaveChanges();
-
-           
         }
 
         public ParameterDto GetParameterById(int parameterId)
@@ -93,7 +81,6 @@ namespace ReasnAPI.Services {
             return parameterDtos;
         }
 
-        
        public List<ParameterDto> GetParametersByFilter(Expression<Func<Parameter, bool>> filter)
         {
             var parameters = _context.Parameters.Where(filter).ToList();
